Moved filter_tags to fix CI fails
