+++ aiarena/api/view_filters.py
@@ -1,7 +1,7 @@
 from django_filters import rest_framework as filters
 
 from aiarena.core.models import Match, Result, Bot, Map, User, Round, MatchParticipation, CompetitionParticipation, Competition
-from aiarena.core.utils import filter_tags
+from aiarena.core.d_utils import filter_tags
 
 
 # Filter for the API views
